chore: move NewMessagePage hooks to top level
Moves `usePlacesWithPaEss` and `useRouteToRouteIDsMap` to the top-level
hooks directory because these are useful beyond just the New PA Message
page.
