+++ assets/js/components/Dashboard/NewPaMessage/hooks.tsx
@@ -1,22 +1,5 @@
 import { useMemo } from "react";
 import { Place } from "Models/place";
-import { useScreenplayContext } from "Hooks/useScreenplayContext";
-import { busRouteIdsAtPlaces } from "../../../util";
-import { BASE_ROUTE_NAME_TO_ROUTE_IDS } from "Constants/constants";
-
-export const usePlacesWithPaEss = () => {
-  const { places } = useScreenplayContext();
-  return useMemo(
-    () =>
-      places
-        .map((place) => ({
-          ...place,
-          screens: place.screens.filter((screen) => screen.type === "pa_ess"),
-        }))
-        .filter((place: Place) => place.screens.length > 0),
-    [places],
-  );
-};
 
 export const usePlacesWithSelectedScreens = (
   places: Place[],
@@ -30,14 +13,4 @@ export const usePlacesWithSelectedScreens = (
       }))
       .filter((place) => place.screens.length > 0);
   }, [places, signIds]);
-};
-
-export const useRouteToRouteIDsMap = (): { [key: string]: string[] } => {
-  const places = usePlacesWithPaEss();
-  return useMemo(() => {
-    return {
-      ...BASE_ROUTE_NAME_TO_ROUTE_IDS,
-      Bus: busRouteIdsAtPlaces(places),
-    };
-  }, [places]);
 };
