+++ assets/js/hooks/usePlacesWithPaEss.ts
@@ -1,0 +1,17 @@
+import { useMemo } from "react";
+import { Place } from "Models/place";
+import { useScreenplayContext } from "Hooks/useScreenplayContext";
+
+export const usePlacesWithPaEss = () => {
+  const { places } = useScreenplayContext();
+  return useMemo(
+    () =>
+      places
+        .map((place) => ({
+          ...place,
+          screens: place.screens.filter((screen) => screen.type === "pa_ess"),
+        }))
+        .filter((place: Place) => place.screens.length > 0),
+    [places],
+  );
+};
